Move escape() to utils to be used by both XML parsers
